% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjba_baixar_cpopg.R
\name{tjba_baixar_cpopg}
\alias{tjba_baixar_cpopg}
\title{Baixa dados processuais de primeira instância do TJBA}
\usage{
tjba_baixar_cpopg(processo, diretorio = ".")
}
\arguments{
\item{processo}{Vetor de processos}

\item{diretorio}{Onde armazenar os dados}
}
\value{
htmls
}
\description{
Baixa dados processuais de primeira instância do TJBA
}
