% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjba_baixar_sentencas.R
\name{tjba_baixar_sentencas}
\alias{tjba_baixar_sentencas}
\title{Baixa dados processuais de primeira instância do TJBA}
\usage{
tjba_baixar_sentencas(processo, url, diretorio = ".")
}
\arguments{
\item{processo}{Vetor de processos}

\item{url}{Vetor de urls}

\item{diretorio}{Onde armazenar os dados}
}
\value{
pdfs
}
\description{
Baixa dados processuais de primeira instância do TJBA
}
