% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjba_ler_movimentacoes.R
\name{tjba_ler_movimentacoes}
\alias{tjba_ler_movimentacoes}
\title{Ler movimentação do TJBA}
\usage{
tjba_ler_movimentacoes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar diretório se não informar arquivos}
}
\value{
tibble
}
\description{
Ler movimentação do TJBA
}
